Famous Paintings/requirements.txt
fsspec==2024.10.0
greenlet==3.1.1
numpy==2.1.3
pandas==2.2.3
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2
